backend/pyproject.toml
[tool.poetry]
name = "bridge-hack"
version = "0.1.0"
description = ""
authors = ["Pedro Binotto <pedrosantibinotto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.45.1"
datasets = "^3.0.1"
torch = "^2.4.1"
scikit-learn = "^1.5.2"
fastapi = {extras = ["standard"], version = "^0.115.0"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
